requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bitsandbytes==0.44.1
bleach==6.2.0
blessed==1.20.0
cachetools==5.5.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.6.0
contourpy==1.3.0
cycler==0.12.1
datasets==2.17.0
debugpy==1.8.12
decorator==5.1.1
deepspeed==0.15.3
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.10.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistral_common==1.4.4
mistune==3.1.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
nvitop==1.4.0
openai==1.53.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outlines==0.0.46
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
ray==2.38.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
sacrebleu==2.5.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.1
tomli==2.2.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.46.1
triton==3.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.31
ultralytics-thop==2.0.11
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vllm==0.6.3.post1
wandb==0.18.5
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.13
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
